Refactor cli-converters, remove converters below version 2.2
